requirements.txt
Flask==2.3.2
flask-marshmallow==0.15.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.0.33
selenium==4.10.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pandas==2.0.3
pluggy==1.0.0
PyJWT==2.7.0
PyMySQL==1.0.3
pytest==7.3.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
tomli==2.0.1
Werkzeug==2.3.4
zipp==3.15.0